InternetSitem/requirements.txt
Flask==3.0.0
qrcode[pil]==8.0
Pillow==10.1.0
gunicorn==21.2.0
